+++ packages/block-library/src/navigation/edit/navigation-link-ui.js
@@ -1,12 +1,5 @@
-/**
- * WordPress dependencies
- */
 import { store as blockEditorStore } from '@wordpress/block-editor';
 import { useDispatch, useSelect } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
 import {
 	LinkUI,
 	updateAttributes,