+++ packages/dataviews/src/components/dataform-controls/radio.tsx
@@ -1,12 +1,5 @@
-/**
- * WordPress dependencies
- */
 import { privateApis, Spinner } from '@wordpress/components';
 import { useCallback } from '@wordpress/element';
-
-/**
- * Internal dependencies
- */
 import type { DataFormControlProps } from '../../types';
 import { unlock } from '../../lock-unlock';
 import getCustomValidity from './utils/get-custom-validity';