+++ packages/editor/src/components/sidebar/dataform-post-summary.js
@@ -1,6 +1,3 @@
-/**
- * WordPress dependencies
- */
 import { __ } from '@wordpress/i18n';
 import { useDispatch, useSelect, useRegistry } from '@wordpress/data';
 import { store as coreDataStore } from '@wordpress/core-data';
@@ -8,10 +5,6 @@ import { DataForm } from '@wordpress/dataviews';
 import { Stack } from '@wordpress/ui';
 import { useMemo } from '@wordpress/element';
 import { useViewConfig } from '@wordpress/views';
-
-/**
- * Internal dependencies
- */
 import PostCardPanel from '../post-card-panel';
 import PluginPostStatusInfo from '../plugin-post-status-info';
 import PostPanelSection from '../post-panel-section';