+++ packages/global-styles-ui/src/font-library/types.ts
@@ -1,6 +1,3 @@
-/**
- * WordPress dependencies
- */
 import type { FontFamilyPreset } from '@wordpress/global-styles-engine';
 import type { FontFamily, FontFace } from '@wordpress/core-data';