+++ packages/global-styles-engine/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Bug Fixes
+
+-   Render block element styles defined only inside responsive viewport states.
+
+## 1.19.0 (2026-07-29)
+
 ### Internal
 
 -   Update `memize` to 2.1.1 ([#80764](https://github.com/WordPress/gutenberg/pull/80764)).